refactor: Move Order-related classes to new package for improved directory structure
