--- service-order/module-core/src/main/java/com/github/msafriends/serviceorder/modulecore/domain/CouponDiscountType.java
@@ -1,6 +1,0 @@
-package com.github.msafriends.serviceorder.modulecore.domain;
-
-public enum CouponDiscountType {
-    PERCENTAGE,
-    FIXED
-}
