+++ service-order/module-core/src/main/java/com/github/msafriends/serviceorder/modulecore/domain/bank/BankProvider.java
@@ -1,4 +1,4 @@
-package com.github.msafriends.serviceorder.modulecore.domain;
+package com.github.msafriends.serviceorder.modulecore.domain.bank;
 
 import jakarta.persistence.*;
 import lombok.AccessLevel;
