+++ service-order/module-core/src/main/java/com/github/msafriends/serviceorder/modulecore/domain/card/CardExpiration.java
@@ -1,4 +1,4 @@
-package com.github.msafriends.serviceorder.modulecore.domain;
+package com.github.msafriends.serviceorder.modulecore.domain.card;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Embeddable;
