+++ service-order/module-core/src/main/java/com/github/msafriends/serviceorder/modulecore/domain/coupon/CouponDiscountType.java
@@ -1,0 +1,6 @@
+package com.github.msafriends.serviceorder.modulecore.domain.coupon;
+
+public enum CouponDiscountType {
+    PERCENTAGE,
+    FIXED
+}
